Fix deployment.keys feature for container backend.
Added a call to self.set_common_state() to populate the keys from the machine
definition to the machine state. Otherwise the keys are not found during
deployment because the default value of the machine state (empty dict) will
be used.

Added another level of bash invocation to the run_command() method of the
container backend. This is needed to allow setting the 'HOME=/root' env
variable together with commands starting with a subshell. Otherwise we get
a systax error from bash because commands like these are not allowed:
'HOME=/root (echo $HOME); echo $HOME'

Redirected stderr of nc to /dev/null to prevent the "Killed by signal 1"
error message of nc.