Use nixos-container run with nc to connect to the containers sshd.
When at least one of the containers systemd-units waits for a
deployment-key-unit the containers init script never returns from
starting systemd inside the container. This leads to a deadlock when
starting such a container by 'nixops start -d <name>' because the
deploymanet kes are cpoied after the container is started.

This change doesn't solve the whole problem but it allows us to
access the ssh daemon on the not yet started container. This is not
possible because also the network setup is finished after a successful
start of the container.

The nixos-container run command is used togeth with nc to cross the
border between container host and container.